+++ maloney-plugins/src/test/java/ch/hsr/maloney/maloney_plugins/authenticode/AuthenticodeCatalogJobTest.java
@@ -1,18 +1,16 @@
-package ch.hsr.maloney.maloney_plugins;
+package ch.hsr.maloney.maloney_plugins.authenticode;
 
-
+import ch.hsr.maloney.maloney_plugins.authenticode.AuthenticodeCatalogJob;
 import ch.hsr.maloney.processing.Job;
 import ch.hsr.maloney.processing.JobCancelledException;
 import ch.hsr.maloney.storage.FakeDataSource;
 import ch.hsr.maloney.storage.FakeMetaDataStore;
 import ch.hsr.maloney.util.Context;
 import ch.hsr.maloney.util.Event;
-import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-import java.io.IOException;
 import java.nio.file.Paths;
 import java.util.List;
 import java.util.UUID;
@@ -20,7 +18,7 @@ import java.util.UUID;
 /**
  * Created by roman on 15.05.17.
  */
-public class AuthenticodePEJobTest {
+public class AuthenticodeCatalogJobTest {
     private FakeMetaDataStore fakeMetaDataStore;
     private FakeDataSource fakeDataSource;
     private Context ctx;
@@ -39,18 +37,18 @@ public class AuthenticodePEJobTest {
 
     @Test
     public void IdentifyPE() {
-        String path = "/media/sf_shared/PE/7z.exe";
+        String path = "/media/sf_shared/PE/VBoxNetLwf.cat";
         UUID id = fakeDataSource.addFile(Paths.get(path));
-        Job job = new AuthenticodePEJob();
+        Job job = new AuthenticodeCatalogJob();
         boolean result = job.shouldRun(ctx, new Event("newFile", "test", id));
         Assert.assertTrue("Job should be runnable", result);
     }
 
     @Test
-    public void ValidatePE() throws JobCancelledException {
-        String path = "/media/sf_shared/PE/bci.dll";
+    public void ValidateCatalogFile() throws JobCancelledException {
+        String path = "/media/sf_shared/PE/VBoxNetLwf.cat";
         UUID id = fakeDataSource.addFile(Paths.get(path));
-        Job job = new AuthenticodePEJob();
+        Job job = new AuthenticodeCatalogJob();
         List<Event> result = job.run(ctx, new Event("newFile", "test", id));
     }
 }
