+++ maloney-plugins/src/test/java/ch/hsr/maloney/maloney_plugins/authenticode/AuthenticodePEJobTest.java
@@ -1,17 +1,17 @@
-package ch.hsr.maloney.maloney_plugins;
+package ch.hsr.maloney.maloney_plugins.authenticode;
 
+
+import ch.hsr.maloney.maloney_plugins.authenticode.AuthenticodePEJob;
 import ch.hsr.maloney.processing.Job;
 import ch.hsr.maloney.processing.JobCancelledException;
 import ch.hsr.maloney.storage.FakeDataSource;
 import ch.hsr.maloney.storage.FakeMetaDataStore;
 import ch.hsr.maloney.util.Context;
 import ch.hsr.maloney.util.Event;
-import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-import java.io.IOException;
 import java.nio.file.Paths;
 import java.util.List;
 import java.util.UUID;
@@ -19,7 +19,7 @@ import java.util.UUID;
 /**
  * Created by roman on 15.05.17.
  */
-public class AuthenticodeCatalogJobTest {
+public class AuthenticodePEJobTest {
     private FakeMetaDataStore fakeMetaDataStore;
     private FakeDataSource fakeDataSource;
     private Context ctx;
@@ -38,18 +38,18 @@ public class AuthenticodeCatalogJobTest {
 
     @Test
     public void IdentifyPE() {
-        String path = "/media/sf_shared/PE/VBoxNetLwf.cat";
+        String path = "/media/sf_shared/PE/7z.exe";
         UUID id = fakeDataSource.addFile(Paths.get(path));
-        Job job = new AuthenticodeCatalogJob();
+        Job job = new AuthenticodePEJob();
         boolean result = job.shouldRun(ctx, new Event("newFile", "test", id));
         Assert.assertTrue("Job should be runnable", result);
     }
 
     @Test
-    public void ValidateCatalogFile() throws JobCancelledException {
-        String path = "/media/sf_shared/PE/VBoxNetLwf.cat";
+    public void ValidatePE() throws JobCancelledException {
+        String path = "/media/sf_shared/PE/bci.dll";
         UUID id = fakeDataSource.addFile(Paths.get(path));
-        Job job = new AuthenticodeCatalogJob();
+        Job job = new AuthenticodePEJob();
         List<Event> result = job.run(ctx, new Event("newFile", "test", id));
     }
 }
